backend/requirements.txt
flask
pinecone-client
sentence-transformers
numpy
